pyproject.toml
[tool.poetry]
name = "cookiebutter"
version = "0.1.0"
description = "Cookiecutter, but all buttered up 🧈 Use cookiecutter with yaml files, or even with python files"
authors = ["Tom Gringauz <tomgrin10@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = ""
repository = "https://github.com/tomgrin10/cookiebutter"
documentation = ""

keywords = ["cookiecutter", "better", "yaml"]

classifiers = [
    'Development Status :: 3 - Alpha',
#    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Developers',
    'Topic :: Utilities',
    'Typing :: Typed'
]


[tool.poetry.scripts]
cookiebutter = 'cookiebutter:wrap_cookiecutter'


[tool.poetry.dependencies]
python = "^3.7"
cookiecutter = "*"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^5.4.2"
pytest = "^4.6.11"


[tool.isort]
profile = "black"


[tool.black]
line-length = 90


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
